requirements.txt
# Second Brain Dependencies
# Core framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# File watching
watchdog>=3.0.0

# Configuration
python-dotenv>=1.0.0

# YAML parsing for frontmatter
PyYAML>=6.0.1

# Type hints (Python 3.9+ compatible)
typing-extensions>=4.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
